Minor fix rectangle area through diagonal and side area unsigned short instead of double for efficiency

+++ rectangle_area_through_diagonal_and_side_area/rectangle_area_through_diagonal_and_side_area.cpp
@@ -4,7 +4,7 @@
 
 using namespace std;
 
-double rectangle_area(double a, double d)
+double rectangle_area(unsigned short a, unsigned short d)
 {
     cout << "Rectangle area = ";
